Refactor: Move movie search redux logic to custom hook
    - Extracted movie search logic from SearchMoviesView into a new custom hook `useMovieSearchRedux`.
    - Simplified the component by delegating state management and data fetching to the custom hook.
    - Improved code readability and reusability by encapsulating logic.
