+++ jobs/Frontend/cristinabulnes_movie-search-app/src/pages/SearchMoviesView.tsx
@@ -1,44 +1,20 @@
-import React, { useCallback, useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { AppDispatch, RootState } from "../redux/store";
-import { fetchMoviesAsync } from "../redux/thunk/movieSearchThunk";
-import { loadMore, setQuery } from "../redux/slices/movieSearchSlice";
-import { useDebounce } from "../hooks/useDebounce";
+import React from "react";
 import Input from "../components/Input/Input";
 import MoviesGrid from "../components/MoviesGrid";
 import Loading from "../components/Loading";
 import ErrorComponent from "../components/Error";
+import { useMovieSearchRedux } from "../hooks/useMovieSearchRedux";
 
 const SearchMoviesView = () => {
-	const handleInputChange = useCallback(
-		(e: React.ChangeEvent<HTMLInputElement>) => {
-			dispatch(setQuery(e.target.value));
-		},
-		[]
-	);
-
-	const dispatch = useDispatch<AppDispatch>();
-	const { movies, isLoading, error, hasMore, query, page } = useSelector(
-		(state: RootState) => state.movieSearch
-	);
-
-	const debouncedQuery = useDebounce(query, 500);
-
-	const fetchMovies = useCallback(async () => {
-		if (debouncedQuery) {
-			dispatch(fetchMoviesAsync({ query: debouncedQuery, page }));
-		}
-	}, [dispatch, debouncedQuery, page]);
-
-	useEffect(() => {
-		if (debouncedQuery) {
-			fetchMovies();
-		}
-	}, [debouncedQuery, page, fetchMovies]);
-
-	const handleLoadMore = () => {
-		dispatch(loadMore());
-	};
+	const {
+		query,
+		handleInputChange,
+		movies,
+		isLoading,
+		error,
+		handleLoadMore,
+		hasMore,
+	} = useMovieSearchRedux();
 
 	return (
 		<>
